chore: move annotation and argument provider classes to new util package
