+++ app/src/test/kotlin/model/UndirectedGraphTest.kt
@@ -4,26 +4,8 @@ import model.abstractGraph.Edge
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.extension.ExtensionContext
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.Arguments
-import org.junit.jupiter.params.provider.ArgumentsProvider
-import org.junit.jupiter.params.provider.ArgumentsSource
-import java.util.stream.Stream
-
-
-@ParameterizedTest(name = "{0}")
-@ArgumentsSource(WeightedAndUnweightedUndirectedGraphsProvider::class)
-@Target(AnnotationTarget.FUNCTION)
-@Retention(AnnotationRetention.RUNTIME)
-private annotation class TestBothUndirectedGraphs
-
-class WeightedAndUnweightedUndirectedGraphsProvider : ArgumentsProvider {
-    override fun provideArguments(context: ExtensionContext?): Stream<Arguments> = Stream.of(
-        Arguments.of(UndirectedGraph<Int>()),
-        Arguments.of(WeightedUndirectedGraph<Int>())
-    )
-}
+import util.annotations.TestAllUndirectedGraphs
+import util.setup
 
 class UndirectedGraphTest {
     private lateinit var graph: UndirectedGraph<Int>
