+++ app/src/test/kotlin/model/WeightedUndirectedGraphTest.kt
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
+import util.setup
 
 class WeightedUndirectedGraphTest {
     private lateinit var graph: WeightedUndirectedGraph<Int>
